pyproject.toml
[project]
name = "sign-id"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
]

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
fixable = ["ALL"]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["signature_model"]
